Renamed namespaces to plural
http://programmers.stackexchange.com/questions/75919/should-package-names-be-singular-or-plural/75929
