+++ library/Swagger/Processors/PartialProcessor.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Swagger\Processor;
+namespace Swagger\Processors;
 
 use Swagger\Logger;
 use Swagger\Parser;
